% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_functions.R
\name{compute_brain_mask}
\alias{compute_brain_mask}
\title{Compute a loose brain mask from functional MRI data using FSL}
\usage{
compute_brain_mask(in_file, log_file = NULL)
}
\arguments{
\item{in_file}{Path to the input 4D NIfTI functional image.}

\item{log_file}{Optional path to a log file for command output.}
}
\value{
File path to the computed binary brain mask (not yet dilated). A dilated version
of the mask is also saved with a \verb{_dil1x} suffix.
}
\description{
Generates a brain mask from a functional image using a modified FSL approach
based on the 98-2 percentile intensity method. This method combines BET skull-stripping
with percentile thresholding and binary dilation to produce a conservative mask.
}
\details{
This function replicates the "98-2" heuristic used in FSL’s featlib.tcl:
it computes the 2nd and 98th percentiles from a skull-stripped mean image and thresholds
at 10\% above the 2nd percentile. A final mask is formed by applying this threshold,
binarizing, and performing one dilation iteration.
}
\keyword{internal}
