% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{get_bids_subjects}
\alias{get_bids_subjects}
\title{Helper function to local all subject directories within a BIDS-compatible folder}
\usage{
get_bids_subjects(bids_root = NULL, full.names = FALSE)
}
\arguments{
\item{bids_root}{The path to a BIDS-compatible folder}

\item{full.names}{If \code{TRUE}, return the absolute paths to the subject folders. If
\code{FALSE}, return paths relative to \code{bids_root}. Default: \code{FALSE}.}
}
\value{
a character vector of subject folders located in \code{bids_root}
}
\description{
Helper function to local all subject directories within a BIDS-compatible folder
}
\keyword{internal}
