% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_nested_values}
\alias{get_nested_values}
\title{Helper function to allow a nested list to be traversed using a key string supporting nesting}
\usage{
get_nested_values(lst, key_strings, sep = "/", simplify = TRUE)
}
\arguments{
\item{lst}{a list to be traversed}

\item{key_strings}{a character vector of keys to traverse the list. Each key string should be
a single string with the keys separated by a separator (default is "/"). For example, "parent/child/grandchild"
would correspond to my_list$parent$child$grandchild}

\item{sep}{a character string to separate the keys in the key strings. Default is "/"}

\item{simplify}{a boolean indicating whether to simplify the output. Default is TRUE}
}
\value{
a named list of values corresponding to the keys in the key strings
}
\description{
Helper function to allow a nested list to be traversed using a key string supporting nesting
}
\keyword{internal}
