% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{get_subject_dirs}
\alias{get_subject_dirs}
\title{Helper function to obtain all subject and session directories from a root folder}
\usage{
get_subject_dirs(
  root = NULL,
  sub_regex = "[0-9]+",
  sub_id_match = "([0-9]+)",
  ses_regex = NULL,
  ses_id_match = "([0-9]+)",
  full.names = FALSE
)
}
\arguments{
\item{root}{The path to a root folder containing subject folders.}

\item{sub_regex}{A regex pattern to match the subject folders. Default: \code{"[0-9]+"}.}

\item{sub_id_match}{A regex pattern to indicate how to extract the subject ID from the subject folder name.
Default: \code{"([0-9]+)"}. The default will extract the first number-like sequence from the folder name.
Note that this can include multiple capturing groups if the ID has multiple parts.}

\item{ses_regex}{A regex pattern to match the session folders. Default: \code{NULL}.
If \code{NULL}, only the subject folders are returned. If not \code{NULL}, the session folders
are returned as well.}

\item{ses_id_match}{A regex pattern to indicate how to extract the session ID from the session folder name.
Default: \code{"([0-9]+)"}. The default will extract the first number-like sequence from the folder name.}

\item{full.names}{If \code{TRUE}, return the absolute paths to the subject folders. If
\code{FALSE}, return paths relative to \code{root}. Default: \code{TRUE}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{sub_id}: The subject ID extracted from the subject folder name.
\item \code{ses_id}: The session ID extracted from the session folder name. If no session folders are found,
this column will be \code{NA}.
\item \code{sub_dir}: The path to the subject folder.
\item \code{ses_dir}: The path to the session folder. If no session folders are found, this column will be \code{NA}.
}
}
\description{
Helper function to obtain all subject and session directories from a root folder
}
\details{
This function is used to find all subject folders within a root folder.
It is used internally by the package to find the subject DICOM and BIDS folders for processing.
The function uses the \code{list.dirs} function to list all directories within the
folder and then filters the directories based on the regex patterns provided.
The function returns a character vector of the subject folders found.

The function also extracts the subject and session IDs from the folder names
using the regex patterns provided. The IDs are extracted using the \code{extract_capturing_groups}
function, which uses the \code{regexec} and \code{regmatches} functions to extract the capturing groups
from the folder names. The function returns a data frame with the subject and session IDs
and the corresponding folder paths.
}
\examples{
  get_subject_dirs(root = "/path/to/root", sub_regex = "[0-9]+", sub_id_match = "([0-9]+)",
    ses_regex = "ses-[0-9]+", ses_id_match = "([0-9]+)", full.names = TRUE)
}
\keyword{internal}
