% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{get_subject_logger}
\alias{get_subject_logger}
\title{This function returns an lgr object corresponding to logging for a single subject directory}
\usage{
get_subject_logger(scfg, sub_id)
}
\arguments{
\item{scfg}{The study configuration object}

\item{sub_dir}{The directory of the target subject's folder within the BIDS structure}
}
\value{
a configured lgr object for logging subject processing messages
}
\description{
This function returns an lgr object corresponding to logging for a single subject directory
}
\keyword{internal}
