% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{load_study}
\alias{load_study}
\title{Load a study configuration from a file}
\usage{
load_study(file = NULL, validate = TRUE)
}
\arguments{
\item{file}{a YAML file containing a valid study configuration}
}
\description{
Load a study configuration from a file
}
