% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{parse_cli_args}
\alias{parse_cli_args}
\title{Parse CLI-style arguments into a nested list using args_to_df()}
\usage{
parse_cli_args(args, sep = "/", type_values = TRUE)
}
\arguments{
\item{args}{A character vector (e.g., from \code{commandArgs(trailingOnly = TRUE)}).}

\item{sep}{A character used to separate nested keys. Default is \code{"/"}.}

\item{type_values}{Logical; whether to attempt to conver right-hand side strings to relevant data
types using \code{type.convert}.}
}
\value{
A nested list of parsed CLI arguments.
}
\description{
This function tokenizes command-line arguments using \code{args_to_df()} and builds a nested list
by interpreting forward slashes in keys (e.g., \code{--a/b=10 11}) as nested structure.
}
\keyword{internal}
