% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{parse_number_base}
\alias{parse_number_base}
\title{helper function to extract numbers from number-like strings}
\usage{
parse_number_base(x)
}
\arguments{
\item{x}{a character vector containing number-like strings}
}
\value{
a numeric vector containing the parsed values
}
\description{
helper function to extract numbers from number-like strings
}
\examples{
  parse_number_base(c("1,234.56", "7,890", "12.34"))
}
\keyword{internal}
