% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_subject.R
\name{process_subject}
\alias{process_subject}
\title{Preprocess a single subject}
\usage{
process_subject(scfg, sub_cfg = NULL, steps = NULL)
}
\arguments{
\item{scfg}{A list of configuration settings}
}
\description{
Preprocess a single subject
}
