% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{prompt_input}
\alias{prompt_input}
\title{Obtain user input from the console}
\usage{
prompt_input(
  prompt = "",
  instruct = NULL,
  type = "character",
  lower = -Inf,
  upper = Inf,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  split = NULL,
  among = NULL,
  required = TRUE,
  uniq = FALSE
)
}
\arguments{
\item{prompt}{The character string to place in front of the user input prompt. For example, "> "}

\item{instruct}{The instructions to display above the prompt.}

\item{lower}{For numeric inputs, the lowest valid value}

\item{upper}{For numeric inputs, the highest valid value}
}
\description{
Obtain user input from the console
}
\keyword{internal}
