% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{read_multiline_input}
\alias{read_multiline_input}
\title{helper function}
\usage{
read_multiline_input(
  instruct = NULL,
  prompt = "> ",
  n_blank = 1,
  collapse = NULL
)
}
\description{
helper function
}
