% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_study.R
\name{run_study}
\alias{run_study}
\title{Run the processing pipeline}
\usage{
run_study(scfg, steps = NULL, prompt = TRUE, debug = FALSE, force = FALSE)
}
\arguments{
\item{scfg}{A list containing the study configuration.}

\item{prompt}{A logical value indicating whether to prompt the user for input on which steps to run.}

\item{force}{A logical value indicating whether to force the execution of all steps, regardless of their current status.}
}
\value{
A logical value indicating whether the processing pipeline was successfully run.
}
\description{
Run the processing pipeline
}
\examples{
\dontrun{
# Assuming you have a valid study configuration list named `study_config`
run_study(study_config, prompt = TRUE, force = FALSE)
}
}
