% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{set_cli_options}
\alias{set_cli_options}
\title{helper function that takes a character vector of CLI arguments and replaces matching old values with
intended new values}
\usage{
set_cli_options(args = NULL, new_values = NULL)
}
\arguments{
\item{args}{a character vector of existing CLI arguments}

\item{new_values}{a character vector of new CLI arguments to be substituted into \code{args}}
}
\value{
a modified character vector of CLI arguments
}
\description{
helper function that takes a character vector of CLI arguments and replaces matching old values with
intended new values
}
\keyword{internal}
