% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{set_nested_values}
\alias{set_nested_values}
\title{Assign values to a nested list using key-value strings}
\usage{
set_nested_values(assignments, sep = "/", lst = NULL, type_values = TRUE)
}
\arguments{
\item{assignments}{A character vector of assignment strings (e.g., \code{"a/b=1"}, \code{"x/y/z=TRUE"}).}

\item{sep}{A character used to separate keys. Default is \code{"/"}.}

\item{lst}{Optional list to update. If \code{NULL}, starts from an empty list.}

\item{type_values}{Logical; whether to attempt to conver right-hand side strings to relevant data
types using \code{type.convert}.}
}
\value{
A nested list with the specified keys and values.
}
\description{
Parses assignments like \code{"a/b/c=10"} and returns \code{list(a = list(b = list(c = 10)))}.
}
\keyword{internal}
