% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_aroma}
\alias{setup_aroma}
\title{Specify the aroma settings}
\usage{
setup_aroma(scfg, fields = NULL)
}
\arguments{
\item{scfg}{a study configuration object, as produced by \code{load_study} or \code{setup_study}}
}
\value{
a modified version of \code{scfg} with \verb{$aroma} populated
}
\description{
Specify the aroma settings
}
\keyword{internal}
