% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_compute_environment}
\alias{setup_compute_environment}
\title{Setup the compute environment for a study}
\usage{
setup_compute_environment(scfg = list(), fields = NULL)
}
\arguments{
\item{scfg}{a study configuration object, as produced by \code{load_study} or \code{setup_study}}
}
\value{
a modified version of \code{scfg} with \verb{$compute_environment} populated
}
\description{
Setup the compute environment for a study
}
\keyword{internal}
