% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postproc.R
\name{setup_confound_calculate}
\alias{setup_confound_calculate}
\title{Specify the confound calculation settings for postprocessing}
\usage{
setup_confound_calculate(scfg = list())
}
\arguments{
\item{scfg}{a study configuration object created by \code{setup_study}}
}
\value{
a modified version of \code{scfg} with the \verb{$postprocess$confound_calculate} field populated
}
\description{
Specify the confound calculation settings for postprocessing
}
\keyword{internal}
