% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_fmriprep}
\alias{setup_fmriprep}
\title{Specify the fmriprep settings}
\usage{
setup_fmriprep(scfg = NULL, fields = NULL)
}
\arguments{
\item{scfg}{a study configuration object, as produced by \code{load_study} or \code{setup_study}}
}
\value{
a modified version of \code{scfg} with \verb{$fmriprep} populated
}
\description{
Specify the fmriprep settings
}
\keyword{internal}
