% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_heudiconv}
\alias{setup_heudiconv}
\title{Specify the heudiconv settings}
\usage{
setup_heudiconv(scfg, prompt_all = FALSE)
}
\arguments{
\item{scfg}{a study configuration object, as produced by \code{load_study} or \code{setup_study}}
}
\value{
a modified version of \code{scfg} with \verb{$heudiconv} populated
}
\description{
Specify the heudiconv settings
}
\keyword{internal}
