% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_mriqc}
\alias{setup_mriqc}
\title{Specify the mriqc settings}
\usage{
setup_mriqc(scfg, fields = NULL)
}
\arguments{
\item{scfg}{a study configuration object, as produced by \code{load_study} or \code{setup_study}}
}
\value{
a modified version of \code{scfg} with \verb{$mriqc} populated
}
\description{
Specify the mriqc settings
}
\keyword{internal}
