% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{setup_postproc_steps}
\alias{setup_postproc_steps}
\title{Specify the postprocessing steps for a study}
\usage{
setup_postproc_steps(scfg = list(), fields = NULL)
}
\arguments{
\item{scfg}{a study configuration object produced by \code{setup_study}}

\item{fields}{a character vector of fields to be prompted for. If \code{NULL}, all fields will be prompted for.}
}
\value{
a modified version of \code{scfg} with the \verb{$postprocess$processing_steps} field populated
}
\description{
Specify the postprocessing steps for a study
}
\details{
This function is used to set up the postprocessing steps for a study. It prompts the user for
the order of the processing steps and whether to apply them. The order of the processing steps is important,
particularly because if we filter certain frequencies from the fMRI data, we must filter any regressors that we
later apply to the data -- that is, confounds and fMRI data must match in frequency content prior to regression.
See Hallquist, Hwang, & Luna (2013) or Lindquist (2019) for details.
}
\keyword{internal}
