% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{setup_postprocess}
\alias{setup_postprocess}
\title{Specify postprocessing setting for a study}
\usage{
setup_postprocess(scfg = list(), fields = NULL)
}
\arguments{
\item{scfg}{a study configuration object produced by \code{setup_study}}

\item{fields}{a character vector of fields to be prompted for. If \code{NULL}, all fields will be prompted for.}
}
\value{
a modified version of \code{scfg} with the \verb{$postprocess} field populated
}
\description{
Specify postprocessing setting for a study
}
\keyword{internal}
