% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postproc.R
\name{setup_postprocessing}
\alias{setup_postprocessing}
\title{Specify postprocessing setting for a study}
\usage{
setup_postprocessing(scfg = list())
}
\arguments{
\item{scfg}{a study configuration object produced by \code{setup_study}}
}
\description{
Specify postprocessing setting for a study
}
