% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postproc.R
\name{setup_spatial_smooth}
\alias{setup_spatial_smooth}
\title{Specify the spatial smoothing settings for postprocessing}
\usage{
setup_spatial_smooth(scfg = list())
}
\arguments{
\item{scfg}{a study configuration object created by \code{setup_study}}
}
\value{
a modified version of \code{scfg} with the \verb{$postprocess$spatial_smooth} field populated
}
\description{
Specify the spatial smoothing settings for postprocessing
}
\keyword{internal}
