% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_study}
\alias{setup_study}
\title{Setup the processing pipeline for a new fMRI study}
\usage{
setup_study(file = NULL)
}
\arguments{
\item{file}{a YAML file specifying the location of an existing configuration file to be loaded}
}
\description{
Setup the processing pipeline for a new fMRI study
}
