% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{summary.bg_study_cfg}
\alias{summary.bg_study_cfg}
\title{summary method for study configuration object}
\usage{
\method{summary}{bg_study_cfg}(x)
}
\description{
summary method for study configuration object
}
