% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{validate_exists}
\alias{validate_exists}
\title{Helper function to check whether a given file or directory exists and, optionally, is readable}
\usage{
validate_exists(
  f,
  description = "",
  directory = FALSE,
  prompt_change = FALSE,
  check_readable = TRUE
)
}
\arguments{
\item{f}{a file or directory to check for existence}

\item{description}{a character string describing what this file is if we are prompted to change it}

\item{directory}{if TRUE, check whether a directory exists. If FALSE (default), check that the file exists}

\item{prompt_change}{if TRUE, if the file/directory exists, ask the user if they wish to change the value. If so, return FALSE}

\item{check_readable}{if TRUE, validation fails (return \code{FALSE}) when the file/directory exists but is not readable}
}
\value{
a boolean (\code{TRUE/FALSE}) indicating whether the file or directory exists and is valid
}
\description{
Helper function to check whether a given file or directory exists and, optionally, is readable
}
\keyword{internal}
