% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_study.R
\name{validate_postprocess_config}
\alias{validate_postprocess_config}
\title{Validate postprocess configuration block}
\usage{
validate_postprocess_config(ppcfg, quiet = FALSE)
}
\arguments{
\item{ppcfg}{a postprocess configuration block}

\item{quiet}{a flag indicating whether to suppress messages}
}
\description{
Validate postprocess configuration block
}
\keyword{internal}
