% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{validate_scfg}
\alias{validate_scfg}
\title{validate the structure of a study configuration object}
\usage{
validate_scfg(scfg = list(), quiet = FALSE)
}
\arguments{
\item{scfg}{a study configuration file as produced by \code{load_study} or \code{setup_study}}
}
\description{
validate the structure of a study configuration object
}
\keyword{internal}
