% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_study.R
\name{validate_study}
\alias{validate_study}
\title{Validate the structure of a study configuration object}
\usage{
validate_study(scfg = list(), quiet = FALSE)
}
\arguments{
\item{scfg}{a study configuration file as produced by \code{load_study} or \code{setup_study}}
}
\description{
Validate the structure of a study configuration object
}
\keyword{internal}
