% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_postprocess.R
\name{setup_apply_aroma}
\alias{setup_apply_aroma}
\title{Specify the intensity normalization settings for postprocessing}
\usage{
setup_apply_aroma(scfg = list(), fields = NULL)
}
\arguments{
\item{scfg}{a study configuration object created by \code{setup_study}}

\item{fields}{a character vector of fields to be prompted for. If \code{NULL}, all fields will be prompted for.}
}
\value{
a modified version of \code{scfg} with the \verb{$postprocess$apply_aroma} field populated
}
\description{
Specify the intensity normalization settings for postprocessing
}
\keyword{internal}
