% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_study.R
\name{setup_study}
\alias{setup_study}
\title{Setup the processing pipeline for a new fMRI study}
\usage{
setup_study(input = NULL, fields = NULL)
}
\arguments{
\item{input}{an existing \code{bg_study_cfg} object to be modified or a string
specifying the location of an existing configuration YAML file to be loaded}

\item{fields}{a character vector of fields to be prompted for. If \code{NULL}, all fields will be prompted for.}
}
\description{
Setup the processing pipeline for a new fMRI study
}
